+++ .vscode/settings.json
@@ -4,5 +4,8 @@
 	"spellright.documentTypes": ["yaml", "git-commit", "markdown"],
 	"typescript.tsdk": "node_modules/typescript/lib",
 	"editor.tabSize": 2,
-	"eslint.enable": true
+	"eslint.enable": true,
+	"editor.codeActionsOnSave": {
+		"source.organizeImports": true
+	}
 }
