+++ publicodes/source/mecanisms/possibility.tsx
@@ -1,7 +1,4 @@
-import React from 'react'
-import { evaluateNode } from '../evaluation'
-import { RuleLinkWithContext } from '../components/RuleLink'
-import { Mecanism } from '../components/mecanisms/common'
+import { registerEvaluationFunction } from '../evaluation'
 
 const evaluate = (cache, situation, parsedRules, node) => {
 	return { ...node }
@@ -10,12 +7,14 @@ const evaluate = (cache, situation, parsedRules, node) => {
 export const mecanismPossibility = (recurse, k, v) => {
 	return {
 		explanation: {},
-		evaluate,
 		jsx: function Synchronisation({ explanation }) {
 			return null
 		},
 		category: 'mecanism',
 		name: 'possibilité',
+		nodeKind: 'possibilité',
 		type: 'possibilité'
 	}
 }
+
+registerEvaluationFunction('possibilité', evaluate)
